+++ src/utils_executer_redirec.c
@@ -25,6 +25,7 @@ t_token_list	*iterate_to_current_pipe(t_data *data)
 		return (data->tokens);
 	return (current);
 }
+
 /**
  * 	@brief	handles redirec and heredoc depending on the types
  */
@@ -52,16 +53,6 @@ static void	handle_types(t_data *data)
 	}
 }
 
-static int	twod_elems(char **cmd_line)
-{
-	int	i;
-
-	i = 0;
-	while (cmd_line[i])
-		i++;
-	return (i * 4);
-}
-
 int	countyy(char **cmd_line)
 {
 	int	i;
@@ -84,69 +75,6 @@ int	countyy(char **cmd_line)
 	return (count);
 }
 
-// FIXME
-// somethings going wrong with grep and awk after a pipe
-char	*remove_redirec(char **cmd_line)
-{
-	char	*cmd;
-	int		i;
-	int		j;
-	int		k;
-
-	cmd = ft_calloc(countyy(cmd_line), sizeof(char));
-	if (cmd == NULL)
-		ft_error(strerror(errno));
-	i = 0;
-	j = 0;
-	while (cmd_line[i])
-	{
-		k = 0;
-		while (cmd_line[i] && (cmd_line[i][k] == '<' || cmd_line[i][k] == '>'))
-			i += 2;
-		if (!cmd_line[i])
-			break ;
-		while (cmd_line[i][k])
-		{
-			cmd[j] = cmd_line[i][k];
-			j++;
-			k++;
-		}
-		cmd[j] = 31;
-		j++;
-		i++;
-	}
-	while (cmd[j])
-	{
-		cmd[j] = '\0';
-		j++;
-	}
-	return (cmd);
-}
-
-void	remove_tmp(t_data *data)
-{
-	char	**cmd_line;
-	int		pid;
-
-	if (access(".tmp.txt", F_OK) == 0)
-	{
-		cmd_line = malloc(sizeof(char *) * 3);
-		cmd_line[0] = "/bin/rm";
-		cmd_line[1] = ".tmp.txt";
-		cmd_line[2] = NULL;
-		pid = fork();
-		if (pid == -1)
-			ft_error(strerror(errno));
-		else if (pid == 0)
-			execve(cmd_line[0], cmd_line, data->env_list);
-		else
-		{
-			wait(NULL);
-			free(cmd_line);
-		}
-	}
-}
-
 /**
  * 	@brief	handles redirections and heredocs
  * 	@param	cmd_line: the old 2d command line
@@ -160,7 +88,7 @@ char	**ft_redirec(char **cmd_line, t_data *data)
 	char	*cmd;
 
 	handle_types(data);
-	cmd = remove_redirec(cmd_line);
+	cmd = remove_redirec(cmd_line, 0, 0);
 	new_cmd_line = ft_split(cmd, 31);
 	free(cmd);
 	remove_tmp(data);
